19. Testing Event Bindings

+++ src/app/2-todos/todos.component.ts
@@ -1,30 +1,30 @@
 import { Component, OnInit } from '@angular/core';
-import { TodoService } from './todo.service'
+import { TodoService } from './todo.service';
 
 @Component({
   selector: 'app-todos',
   templateUrl: './todos.component.html',
-  styleUrls: ['./todos.component.css']
+  styleUrls: ['./todos.component.css'],
 })
 export class TodosComponent implements OnInit {
-  todos: any[] = [];
-  message; 
+  todos: Array<any> = [];
+  message: any;
 
   constructor(private service: TodoService) {}
 
-  ngOnInit() { 
-    this.service.getTodos().subscribe(t => this.todos = t);
+  ngOnInit() {
+    this.service.getTodos().subscribe((t) => (this.todos = t));
   }
 
-  add() { 
+  add() {
     var newTodo = { title: '... ' };
     this.service.add(newTodo).subscribe(
-      t => this.todos.push(t),
-      err => this.message = err);
+      (t: any) => this.todos.push(t),
+      (err) => (this.message = err)
+    );
   }
 
-  delete(id) {
-    if (confirm('Are you sure?'))
-      this.service.delete(id).subscribe();
-  } 
+  delete(id: any) {
+    if (confirm('Are you sure?')) this.service.delete(id).subscribe();
+  }
 }
